+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   ([#121](https://github.com/signalfx/splunk-otel-js/pull/121))
 - Rename `SPLUNK_CONTEXT_SERVER_TIMING_ENABLED`
   ([#149](https://github.com/signalfx/splunk-otel-js/pull/149))
+- Support for Synthetic Run identification
 
 ## 0.8.0 (04-15-2021)