+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   ([#121](https://github.com/signalfx/splunk-otel-js/pull/121))
 - Rename `SPLUNK_CONTEXT_SERVER_TIMING_ENABLED`
   ([#149](https://github.com/signalfx/splunk-otel-js/pull/149))
+- Upgrade to OpenTelemetry SDK 0.22.0, API 1.0.0.
+  ([#153](https://github.com/signalfx/splunk-otel-js/pull/153))
 
 ## 0.8.0 (04-15-2021)